+++ main.h
@@ -3,6 +3,6 @@
 
 #define MATRIX_SIZE_PER_BLOCK 32
 
-__global__ void convolution(unsigned char *matrix, int matrix_width, int matrix_height, int *kernel, int kernel_size);
+__global__ void convolution(unsigned char *matrix, int matrix_width, int matrix_height, float *kernel, int kernel_size);
 
 #endif // MAIN_H
